Adding local chat participants state + updating scopes required
