+++ samples/apps/copilot-chat-app/webapp/src/Constants.ts
@@ -17,7 +17,7 @@ export const Constants = {
         },
         semanticKernelScopes: ['openid', 'offline_access', 'profile'],
         // MS Graph scopes required for loading user information
-        msGraphAppScopes: ['User.Read'],
+        msGraphAppScopes: ['User.ReadBasic.All'],
     },
     bot: {
         profile: {
