+++ samples/apps/copilot-chat-app/webapp/src/components/chat/ChatWindow.tsx
@@ -25,13 +25,11 @@ import { Dismiss16Regular, Edit24Filled, EditRegular } from '@fluentui/react-ico
 import React, { useEffect, useState } from 'react';
 import { AuthHelper } from '../../libs/auth/AuthHelper';
 import { AlertType } from '../../libs/models/AlertType';
-import { IChatUser } from '../../libs/models/ChatUser';
 import { ChatService } from '../../libs/services/ChatService';
-import { useGraph } from '../../libs/useGraph';
 import { useAppDispatch, useAppSelector } from '../../redux/app/hooks';
 import { RootState } from '../../redux/app/store';
 import { addAlert, removeAlert } from '../../redux/features/app/appSlice';
-import { editConversationTitle, setUsersLoaded } from '../../redux/features/conversations/conversationsSlice';
+import { editConversationTitle } from '../../redux/features/conversations/conversationsSlice';
 import { ChatResourceList } from './ChatResourceList';
 import { ChatRoom } from './ChatRoom';
 import { ParticipantsList } from './controls/ParticipantsList';
@@ -101,27 +99,14 @@ export const ChatWindow: React.FC = () => {
     const classes = useClasses();
     const dispatch = useAppDispatch();
     const { instance, inProgress } = useMsal();
-    const msGraph = useGraph();
     const chatService = new ChatService(process.env.REACT_APP_BACKEND_URI as string);
     const { alerts } = useAppSelector((state: RootState) => state.app);
 
     const { conversations, selectedId } = useAppSelector((state: RootState) => state.conversations);
     const chatName = conversations[selectedId].title;
-    const chatUsers = conversations[selectedId].users;
 
     const [title = '', setTitle] = useState<string | undefined>(selectedId);
     const [isEditing, setIsEditing] = useState<boolean>(false);
-
-    useEffect(() => {
-        if (!conversations[selectedId].userDataLoaded) {
-            void msGraph.loadUsers(chatUsers.map((user: IChatUser) => user.id)).then(() => {
-                dispatch(setUsersLoaded(selectedId));
-            });
-        }
-
-        // Limiting dependencies or else this effect is triggered on all user typing events or new chat messages
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [selectedId, chatUsers.length]);
 
     const onSave = async () => {
         if (chatName !== title) {
