+++ samples/apps/copilot-chat-app/webapp/src/components/chat/controls/ParticipantsList.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-import { FC } from 'react';
+import { FC, useEffect, useState } from 'react';
 
 import {
     Avatar,
@@ -14,8 +14,11 @@ import {
     tokens,
 } from '@fluentui/react-components';
 import { IChatUser } from '../../../libs/models/ChatUser';
-import { useAppSelector } from '../../../redux/app/hooks';
+import { useGraph } from '../../../libs/useGraph';
+import { useAppDispatch, useAppSelector } from '../../../redux/app/hooks';
 import { RootState } from '../../../redux/app/store';
+import { setUsersLoaded } from '../../../redux/features/conversations/conversationsSlice';
+import { Users } from '../../../redux/features/users/UsersState';
 import { ScrollBarStyles } from '../../../styles';
 
 interface ShareBotMenuProps {
@@ -42,7 +45,27 @@ const useClasses = makeStyles({
 
 export const ParticipantsList: FC<ShareBotMenuProps> = ({ participants }) => {
     const classes = useClasses();
+    const dispatch = useAppDispatch();
     const { users } = useAppSelector((state: RootState) => state.users);
+    const { conversations, selectedId } = useAppSelector((state: RootState) => state.conversations);
+    const msGraph = useGraph();
+
+    const chatUsers = conversations[selectedId].users;
+    const [loadedParticipants, setLoadedParticipants] = useState(mapParticipantsNames(participants, users));
+
+    useEffect(() => {
+        if (!conversations[selectedId].userDataLoaded) {
+            void msGraph.loadUsers(chatUsers.map((user: IChatUser) => user.id)).then((loadedUsers?: Users) => {
+                if (loadedUsers) {
+                    dispatch(setUsersLoaded(selectedId));
+                    setLoadedParticipants(mapParticipantsNames(participants, loadedUsers));
+                }
+            });
+        }
+
+        // Limiting dependencies or else this effect is triggered on all user typing events or new chat messages
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [selectedId, chatUsers.length]);
 
     return (
         <Popover positioning={'below-end'} size="small">
@@ -53,14 +76,10 @@ export const ParticipantsList: FC<ShareBotMenuProps> = ({ participants }) => {
             </PopoverTrigger>
             <PopoverSurface className={classes.root}>
                 <div className={classes.list}>
-                    {participants.map((participant) => (
+                    {loadedParticipants.map((participant) => (
                         <Persona
                             textAlignment="center"
-                            name={
-                                // falsy lint check, users[userId] can be null if user data hasn't been loaded
-                                // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-                                users[participant.id.split('.')[0]]?.displayName ?? participant.id
-                            }
+                            name={participant.name}
                             key={participant.id}
                             avatar={{ color: 'colorful' }}
                         />
@@ -69,4 +88,13 @@ export const ParticipantsList: FC<ShareBotMenuProps> = ({ participants }) => {
             </PopoverSurface>
         </Popover>
     );
+};
+
+const mapParticipantsNames = (participants: IChatUser[], usersData: Users) => {
+    return participants.map((participant) => ({
+        ...participant,
+        // falsy lint check, users[userId] can be null if user data hasn't been loaded
+        // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+        name: usersData[participant.id.split('.')[0]]?.displayName ?? participant.id,
+    }));
 };
